Use the normal ci api github uses
Github confusingly calls this 'status' and it is attached to commit but
not PR. Which is actually good, but still odd. We have to do some
searching around in the list of commits to find the head commit and vote
on that. This is all done without OAUTH for now, which means the pcci
bot will need push access to do this kind of voting.