Allow us to test each module against multiple nodesets
Ideally this will pull the config from .pcci.yml file. If that doesn't
exist (which it doesn't right now) it will default to the current
behaviour of only testing ubuntu trusty and centos 7.

This also pulls out the nodeset definitions into proper yml files rather
that the current heredoc strings in pcci-configuration.

The intention of this PR is that it will allow us to explode the matrix
of possible test options.