Try/except around repo renames
Unfortunately, we rename repos a lot. Github handles redirects for most
use cases but not when querying repos via the api. As a result, a lot of
pcci explodes when something gets renamed. This change makes it so that
follow_pull_requests.py can skip a broken repo and continue. Note that
someone will need to be watching for this in the logs since this failure
will be more subtle and harder to notice.

https://github.com/voxpupuli/pcci/issues/61